+++ django_socio_grpc/permissions.py
@@ -3,8 +3,8 @@ from typing import TYPE_CHECKING
 from rest_framework.permissions import BasePermission
 
 if TYPE_CHECKING:
-    from django_socio_grpc.request_transformer.grpc_socio_proxy_context import (
-        GRPCSocioProxyContext,
+    from django_socio_grpc.request_transformer.grpc_internal_proxy import (
+        GRPCInternalProxyContext,
     )
     from django_socio_grpc.services import Service
 
@@ -14,8 +14,10 @@ class GRPCActionBasePermission(BasePermission):
     Base class from which all GRPC action permissions should inherit.
     """
 
-    def has_permission(self, context: "GRPCSocioProxyContext", service: "Service"):
+    def has_permission(self, context: "GRPCInternalProxyContext", service: "Service"):
         return True
 
-    def has_object_permission(self, context: "GRPCSocioProxyContext", service: "Service", obj):
+    def has_object_permission(
+        self, context: "GRPCInternalProxyContext", service: "Service", obj
+    ):
         return True