+++ django_socio_grpc/request_transformer/__init_.py
@@ -1,3 +1,0 @@
-from .grpc_socio_proxy_context import GRPCSocioProxyContext
-
-__all__ = ["GRPCSocioProxyContext"]